Complete Overhaul
Absolutely every line of code we had was deleted. It is for the better,
because the code is actually much farther into the game now that this
has happened. There is a main menu state, a game state, and we also have
a package dedicated to the elevator.

Also, there is a short explanation on how Slick2D works inside Elevator
Project.