Anna did work!
I have no idea how to use Git, so let's hope I'm doing this right.
Corrin was kind and drew me buttons for the elevators, so they have now
replaced the panel that was there before. I also changed a few numbers
to prevent some of the text from drifting to the left. The information
at the bottom and the elevator displays were both moving.